requirements.txt
asgiref==3.2.10
atomicwrites==1.4.0
attrs==19.3.0
certifi==2020.4.5.2
chardet==3.0.4
colorama==0.4.3
Django==3.0.7
django-cors-headers==3.4.0
django-filter==2.3.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
idna==2.9
more-itertools==8.4.0
packaging==20.4
pluggy==0.13.1
py==1.8.2
PyJWT==1.7.1
pyparsing==2.4.7
pytest==5.4.3
pytest-django==3.9.0
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
wcwidth==0.2.4
